refactor: relative imports, abstracted Review type and added custom ShareIcon
