+++ src/common/components/CustomIcon/ShareIcon.tsx
@@ -1,0 +1,14 @@
+import { CustomIcon, type CustomIconProps } from "./CustomIcon";
+
+export const ShareIcon = (props: CustomIconProps) => {
+  return (
+    <CustomIcon fill="none" {...props}>
+      <path
+        fill-rule="evenodd"
+        clip-rule="evenodd"
+        d="M14.6303 3.36593C14.3171 3.02334 13.8256 2.90831 13.3928 3.07631C12.9601 3.24431 12.675 3.66084 12.675 4.12504V6.99984C10.7822 7.20245 8.40629 7.925 6.35401 9.54117C3.81872 11.5377 1.87506 14.824 1.87506 19.875C1.87506 20.4963 2.37874 21 3.00006 21C3.62137 21 4.12505 20.4963 4.12505 19.875C4.12505 19.8238 4.23198 19.1448 5.66824 18.4176C6.94155 17.7728 9.12218 17.1819 12.675 17.0785V19.875C12.675 20.3392 12.9601 20.7557 13.3928 20.9237C13.8256 21.0917 14.3171 20.9767 14.6303 20.6341L21.8302 12.7591C22.2232 12.3293 22.2232 11.6707 21.8302 11.2409L14.6303 3.36593Z"
+        fill="currentColor"
+      />
+    </CustomIcon>
+  );
+};
