+++ src/common/components/CustomIcon/index.ts
@@ -25,3 +25,4 @@ export * from "./ThumbUpFilledIcon";
 export * from "./WarningCircleIcon";
 export * from "./XCloseIcon";
 export * from "./SchoolIcon";
+export * from "./ShareIcon";
