+++ src/common/components/ReviewItem/ReviewBody.tsx
@@ -1,6 +1,6 @@
 import { Button } from "@/common/components/Button";
 import { reviewItemTheme, type ReviewItemVariants } from "./ReviewItem.theme";
-import { type Review } from "./ReviewItem";
+import { type Review } from "@/common/types";
 
 export type ReviewBodyProps = ReviewItemVariants & {
   review: Review;
