+++ src/common/components/ReviewItem/index.ts
@@ -1,2 +1,6 @@
 export * from "./ReviewItem";
 export * from "./ReviewItem.theme";
+export * from "./ReviewHeader";
+export * from "./ReviewBody";
+export * from "./ProfileReviewer";
+export * from "./ProfileSchool";
