+++ src/common/components/ReviewModal/ReviewModal.theme.ts
@@ -14,16 +14,23 @@ export const reviewModalTheme = tv(
         "border-surface-base",
         "hover:border-border-elevated",
         "cursor-pointer",
-        "w-[45rem]",
+        "w-full",
+        "max-w-[45rem]",
         "gap-4",
       ],
       modalTrigger: ["w-[45rem]"],
-      modalContent: ["min-w-[600px]"],
+      modalContent: ["w-full", "mx-10"],
       modalBody: ["space-y-8"],
       usernameAndTimestampWrapper: ["space-x-2", "pb-[18px]"],
       username: ["text-lg", "font-medium"],
       body: ["pb-[16px]"],
-      likeAndShareWrapper: ["flex", "gap-x-6"],
+      likeAndShareWrapper: [
+        "flex",
+        "gap-x-6",
+        "items-center",
+        "w-fit",
+        "justify-between",
+      ],
       likeWrapper: [
         "flex",
         "w-fit",
@@ -31,9 +38,9 @@ export const reviewModalTheme = tv(
         "justify-center",
         "gap-x-1",
       ],
-      likeIcon: ["h-fit", "w-[18px]"],
+      // likeIcon: ["h-fit", "w-[18px]"],
       shareWrapper: ["flex", "w-fit", "items-center", "justify-center"],
-      shareIcon: ["h-fit", "w-[18px]"],
+      // shareIcon: ["h-fit", "w-[18px]"],
       seeMoreWrapper: ["w-full", "border-t", "border-border-default", "pt-3"],
       seeMoreLink: ["flex", "h-10", "items-center", "text-primary-default"],
     },
