+++ src/common/types/index.ts
@@ -1,1 +1,2 @@
 export * from "./system";
+export * from "./review";
